Fixing delete() in CollectionDetail

+++ store/views.py
@@ -30,12 +30,12 @@ class CollectionList(ListCreateAPIView):
     serializer_class = CollectionSerializer
 
 class CollectionDetail(RetrieveUpdateDestroyAPIView):
-    queryset = Collection.objects.all()
+    queryset = Collection.objects.annotate(products_count=Count('products'))
     serializer_class = CollectionSerializer
 
     def delete(self, request, pk):
         collection = get_object_or_404(Collection, pk=pk)
-        # if collection.product_count > 0:
-        #     return Response({'errors': 'Collection cannot be delete with associated product_count'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        if collection.products.count() > 0:
+            return Response({'errors': 'Collection cannot be delete with associated product_count'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
